Add global Commitlint hook
Adds a global hook to lint all my Git commit messages with Commitlint
(https://commitlint.js.org/). This was acheived using Git's
core.hooksPath. See https://stackoverflow.com/a/37293198 for details.
The path I choose for hooks was ~/.config/git/hooks.

Commitlint's rules were set up according to Chris Beam's suggested rules
for Git commit messages.

There was an old and buggy pre-commit hook I was working on a while back
to avaid commiting merge conflicted files. I removed it in order to
avoid it running on commits. It'd be worth bringing it back in the
future.

People have reported issues with core.hooksPath and Husky.
(See https://www.jakewiesler.com/blog/husky-and-global-git-hooks/)
Some solutions to this appear to be to alternatively use
init.templateDir or to locally reconfigure core.hooksPath on affected
repos. I think I'll try the latter approach when I run into it; maybe
it'd be worth making a shorthand in the future (e.g., `git
configure-husky`).